Refactoring TypeFlagTooltip
