+++ app/components/TypeFlagTooltip/List.js
@@ -1,0 +1,2 @@
+import styled from 'styled-components';
+export default styled.ul`margin: 0; padding: 0;`;
