+++ app/containers/TypeFlagTooltip/tests/index.test.js
